streamlit_app/requirements.txt
# requirements.txt (Versão Final e Completa)
streamlit
pandas
numpy
yfinance
plotly
requests
python-bcb
sqlalchemy
psycopg2-binary
boto3
streamlit-autorefresh